Tidying, implementing and bug fixing
Moved some functions around, implemented some of the recommendations from SourceryAI and fixed an issue where if a game is added as a favourite when there is nothing in the library, it just would be not selected.

New issue however: need to figure out a way to auto select the game that was added when it is also added as a favourite.